// Client/src/Swagger/models/CategoriesResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Shift
 * Shift Server API documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CategoriesResponse
 */
export interface CategoriesResponse {
    /**
     * 
     * @type {Array<string>}
     * @memberof CategoriesResponse
     */
    categories: Array<string>;
}

export function CategoriesResponseFromJSON(json: any): CategoriesResponse {
    return CategoriesResponseFromJSONTyped(json, false);
}

export function CategoriesResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CategoriesResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'categories': json['categories'],
    };
}

export function CategoriesResponseToJSON(value?: CategoriesResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'categories': value.categories,
    };
}